# docs/openapi.yaml
openapi: 3.0.1
info:
  title: ldap2azure
  description: 'The following describes the API of ldap2azure. The API is used by the frontend, but can also be used with API key authentication from a backend, e.g. for monitoring.'
  license:
    name: GPL-3.0 License
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
  - url: https://localhost:8080/api
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate against ldap2azure using an username and password to gain an session cookie for further actions.
      description: This endpoint is only intended for use in a frontend. To authorize access from another application, please use an ApiKey instead.
      operationId: authLoginPassword
      requestBody:
        description: Login credentials to authenticate against the backend.
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        200:
          description: >
            Successfully authenticated. The response contains your own ApiUser objects as well as
            the session key in a cookie named `cdsess`. You need to include this cookie in all subsequent requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUser'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: cdsess=abcde12345;Version=1;Domain=localhost;Path=/
        401:
          description: Unauthenticated. The given credentials are not valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: invalid_credentials
  /auth/ack:
    get:
      tags:
        - auth
      summary: Prevent a session cookie from expiring.
      description: This endpoint is only intended for use in a frontend. Regularly calling this endpoint prevents a session cookie from expiring after 30 minutes without action.
      operationId: authAck
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: {}
                example: []
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
  /auth/logout:
    get:
      tags:
        - auth
      summary: Invalidate the current session cookie.
      description: This endpoint is only intended for use in a frontend.
      operationId: authLogout
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: {}
                example: []
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
  /auth/api-key:
    post:
      tags:
        - auth
      summary: Creates a new ApiKey.
      description: This endpoint can be used to create static ApiKeys for use in other backends.
      operationId: authCreateKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreateRequest"
      responses:
        200:
          description: Api key successfully created. The secret can be retrieved using another endpoint below.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
      security:
        - cookieAuth: []
        - api_key: []
    get:
      tags:
        - auth
      summary: Get a list of all ApiKeys stored in the database.
      operationId: authGetApiKeys
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: []
        - api_key: []
  /auth/api-key/{id}:
    get:
      tags:
        - auth
      summary: Get a single ApiKey object from the database.
      operationId: authGetApiKey
      parameters:
        - name: id
          description: Id of the ApiKey that should be retreived.
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
      responses:
        200:
          description: Request successfull. The ApiKey with the given id is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        404:
          description: The ApiKey with the given id could not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: apikey_not_existing
      security:
        - cookieAuth: []
        - api_key: []
    delete:
      tags:
        - auth
      summary: Delete an ApiKey from the database. This invalidates the key immediately.
      operationId: authDeleteApiKey
      parameters:
        - name: id
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
          description: Id of the key that should be deleted.
      responses:
        200:
          description: Key sucessfully deleted.
          content:
            application/json:
              schema:
                type: array
                items: {}
                example: []
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: The ApiKey with the given id could not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: apikey_not_existing
      security:
        - cookieAuth: []
        - api_key: []
  /auth/api-key/{id}/secret:
    get:
      tags:
        - auth
      summary: Get the secret for the given ApiKey from the database.
      description: This returns the secret you have to use in the `Authorization` Header to authenticate a backend application.
      operationId: authGetApiKeySecret
      parameters:
        - name: id
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
          description: Id of the ApiKey for which the secret should be fetched.
      responses:
        200:
          description: Secret sucessfully retreived.
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                    example: qKeAeZ/YWobktAfjl?A372xilVC2.=fSjR2.H7m+3dRO*Q%/l5%_3C%$?F#cTFaA
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: The ApiKey with the given id could not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: apikey_not_existing
      security:
        - cookieAuth: []
        - api_key: []
  /auth/api-key/{id}/permission:
    put:
      tags:
        - auth
      summary: Update the permission of the ApiKey with the given id.
      operationId: authUpdateApiKeyPermissionById
      parameters:
        - name: id
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
          description: Id of the ApiKey for which the permission should be updated.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionUpdateRequest"
      responses:
        200:
          description: Permission sucessfully changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        400:
          description: You are trying to change your own ApiKeys permission, which is prevented by the backend.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "cant_change_own_permission"
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: The ApiKey with the given id could not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: apikey_not_existing
      security:
        - cookieAuth: [ ]
        - api_key: [ ]
  /auth/api-user:
    post:
      tags:
        - auth
      summary: Creates a new ApiUser.
      description: This endpoint can be used to create new ApiUsers.
      operationId: authCreateApiUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiUserCreateRequest"
      responses:
        200:
          description: ApiUser successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUser'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
      security:
        - cookieAuth: [ ]
        - api_key: [ ]
    get:
      tags:
        - auth
      summary: Get an list of all currently existing ApiUsers.
      description: This returns an list containing all ApiUsers currently stored in the database.
      operationId: authGetApiUsers
      responses:
        200:
          description: List of ApiUsers sucessfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiUser'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: [ ]
        - api_key: [ ]
  /auth/api-user/{id}:
    get:
      tags:
        - auth
      summary: Get a single ApiUser object from the database.
      operationId: authGetApiUser
      parameters:
        - name: id
          description: Id of the ApiUser that should be retreived.
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
      responses:
        200:
          description: Request successfull. The ApiUser with the given id is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUser'
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        404:
          description: Not found. The ApiUser with the given id not existing in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: apiuser_not_existing
      security:
        - cookieAuth: [ ]
        - api_key: [ ]
    delete:
      tags:
        - auth
      summary: Delete an ApiUser from the database. This also invalidates all of it's child ApiSessions imediatly.
      operationId: authDeleteApiUser
      parameters:
        - name: id
          schema:
            type: string
            example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
          in: path
          required: true
          description: Id of the ApiUser that should be deleted.
      responses:
        200:
          description: ApiUser sucessfully deleted.
          content:
            application/json:
              schema:
                type: array
                items: { }
                example: [ ]
        400:
          description: You are trying delete your own user, which is prevented by the backend.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "cant_delete_own_user"
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
        403:
          description: You have insufficient permissions. This results from an read-only permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "insufficient_permissions"
        404:
          description: ApiUser with the given id could not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: apiuser_not_existing
      security:
        - cookieAuth: [ ]
        - api_key: [ ]
  /auth/api-user/password:
    put:
      tags:
        - auth
      summary: Resets the password of the currently logged in ApiUser.
      description: Update the password of the currently logged in ApiUser. The request will fail if you use an ApiKey for authentication.
      operationId: authUpdateApiUserPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        200:
          description: Password sucessfully changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiUser"
        400:
          description: You are using an ApiKey for authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
              example:
                error: "only_user_sessions_allowed"
        401:
          description: You are not authenticated or your authentication secret is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleLineError'
      security:
        - cookieAuth: [ ]
        - api_key: [ ]
...
components:
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "VerySecurePassword1234"
...
    ApiKey:
      type: object
      properties:
        _id:
          type: string
          example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
        name:
          type: string
          example: ExampleKey
        lastUsed:
          type: string
          format: date-time
        permission:
          type: string
          enum:
            - "read"
            - "read_write"
          example: "read_write"
    ApiKeyCreateRequest:
      type: object
      properties:
        keyName:
          type: string
          example: "ExampleKey"
        permission:
          type: string
          enum:
            - "read"
            - "read_write"
          example: "read_write"
    ApiUser:
      type: object
      properties:
        _id:
          type: string
          example: cb78467f-6dbd-4563-ac01-0fdc9ca4c194
        username:
          type: string
          example: "admin"
        lastLoginTime:
          type: string
          format: date-time
        permission:
          type: string
          enum:
            - "read"
            - "read_write"
          example: "read_write"
    ApiUserCreateRequest:
      type: object
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "VerySecurePassword1234"
        permission:
          type: string
          enum:
            - "read"
            - "read_write"
          example: "read_write"
    PasswordChangeRequest:
      type: object
      properties:
        password:
          type: string
          example: "VerySecurePassword1234"
    PermissionUpdateRequest:
      type: object
      properties:
        permission:
          type: string
          enum:
            - "read"
            - "read_write"
          example: "read_write"
...
    SingleLineError:
      type: object
      properties:
        error:
          type: string
          example: not_authenticated
...
  securitySchemes:
    api_key:
      type: apiKey
      name: X-API-Key
      in: header
    cookieAuth:
      type: apiKey
      in: cookie
      name: cdsess